xmutil: Update apt search command
Remove the grep expression from the apt search command in the getpkgs
subcommand. This way the package description is printed in addition to the
package name. Currently the package description is printed if it includes the
board name but for other packages, the description is skipped. Also, the grep
filter removes newlines and colored output which are useful for better
readability. By removing the grep command, the following status messages are
printed in addition to the package list:

Sorting... Done
Full Text Search... Done

These are useful indicators for the user to understand the progress of the
search query.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>