xmutil: Improve error handling for ipmi-fru calls
Print output of ipmi-fru as part of boardid command, even if the ipmi-fru
command returns an error. Raise an exception in the product_name function
if ipmi-fru returns an error. Catch the exception in the getpkgs command and
print a more meaningful error message. In this case, we don't want to print the
ipmi-fru output as it will be confusing to the user.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>