xmutil: Pass --ignore-errors switch when calling getpkgs
The getpkgs command fails when run on a board with incorrect EEPROM programming
(multirecord checksum error). Hence, pass the --ignore-errors switch so the
function call doesn't terminate on error. The error typically occurs in the last
FRU record but the getpkgs command only needs to read the product name record to
compose the package search string.

Signed-off-by: Christian Kohn <chris.kohn@amd.com>